feat: get API 구현 완료

+++ src/main/java/capstone/walkreen/dto/MissionResponse.java
@@ -22,7 +22,7 @@ public class MissionResponse {
 
     private String contents;
 
-    private Long reward;
+    private Integer reward;
 
     private Long people;
